requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-Login==0.6.2
Werkzeug==2.3.7
SQLAlchemy==2.0.20
python-dotenv==1.0.0
PyJWT==2.8.0 